Refactored v3 code so that v3_01 becomes v3, v3_01 is deleted since no longer necessary
