+++ nodes.py
@@ -2293,7 +2293,6 @@ def init_builtin_extra_nodes():
         "nodes_string.py",
         "nodes_camera_trajectory.py",
         "nodes_v3_test.py",
-        "nodes_v3_01_test.py",
         "nodes_v1_test.py",
     ]
